feat(replays): close the v2 search gap and move dashboard helpers internal
Groundwork for retiring the v1 `sessionReplays` group. Two changes, both
prerequisites for migrating the remaining adapters onto v2.

**`visitor_id` on v2 search.** v1 `ListReplaysRequest` can filter by visitor and
`V2SessionReplaySearchParams` could not, so the "Sessions from visitor" link —
the cross-subdomain walk from an anonymous marketing session to the identified
product sessions of the same browser — had no v2 equivalent to migrate onto.

This is contract-only: `SessionReplaysListOpts.visitorId` and its WHERE clause
already exist in `sessionReplaysListQuery`, and v1 and v2 are two thin HTTP
wrappers over that one query. The handler just never spread the field in. Adds
the three query-engine cases that were missing — the filter was previously
covered only incidentally, under a `userId`-named describe block.

**`facets` and `traceSummaries` move to the internal tier.** Facet counts shape
the replays filter sidebar and trace summaries shape a session's timeline; both
follow what those views render, which is why `docs/http-api-migration.md` marks
them "do not lift" to `/v2`. They now serve from `MapleInternalApi` at
`/internal/session-replays`, session-only and absent from `/docs`, alongside the
query-engine group.

Both are self-contained — `sessionReplaysFacetsQuery` and
`sessionTraceSummariesQuery`, no coupling to the search path — so this is a lift,
not a rewrite. The v1 group keeps `list`, `get`, `for-trace` and `transcript`;
those move to v2 next, and the group goes away with them.
